Moved BMP image creator code to glcd.cpp
